Unify look of VerifiedBadge component across UI
