+++ app/javascript/mastodon/components/badge/styles.module.scss
@@ -10,7 +10,7 @@
 
   > svg {
     width: auto;
-    height: 15px;
+    height: 17px;
     fill: currentColor;
     opacity: 0.85;
   }
@@ -41,7 +41,6 @@
 
 .success {
   background-color: var(--color-bg-success-softest);
-  color: var(--color-text-success);
 }
 
 .warning {
